Move `PATH` manipulation logic to methods of `Image` and `System` and use envoy crate for cleaner `PATH` hacking.
