+++ crates/notion-core/Cargo.toml
@@ -29,3 +29,4 @@ winfolder = "0.1"
 tempfile = "3.0.2"
 os_info = { "git" = "https://github.com/DarkEld3r/os_info", "rev" = "912a7941cd797e9fca83f4fe858ed684baa5fc40" }
 detect-indent = { "git" = "https://github.com/stefanpenner/detect-indent-rs", "branch" = "master" }
+envoy = "0.1.2"
